changed: Removed self-built boost. Use the libs from MinGW Distro instead. removed: Removed self-made Pointer classes. Use std pointers supported by C++11 instead. changed: Make String no longer a descendant of std::wstring. added: Added many utility function to String.

+++ components/common/Object.cpp
@@ -10,7 +10,7 @@ namespace viu2x {
         // Nothing needs to be done.
     }
 
-    uint32_t Object::getHash() {
-        return (uint32_t)this;
+    uint64_t Object::getHash() {
+        return (uint64_t)this;
     }
 }
